Add files via upload
Tableau Dashboard

1st Worksheet

We have first shown a summary of the whole data set and bifurcated it list wise to understand the no. of apps (Top paid, Top Free)

2nd Worksheet

Top Categories:We have given the overview of the top 10 categories ("Games," "Photo & Video, and etc) with the filter of list(Top paid, Top Free)
We can derive from this that “games” is the highest category in both list (Top paid, Top Free) where as “social media” is the highest in “free”and “photos and videos” is the highest in “paid”

3rd Worksheet

Price Trends: We have compared the free vs. paid apps over time, using the "price" and "rank" columns .We have created a calculated field “App Type” with this parameter (IF [price] = 0 THEN "Free" ELSE "Paid" END) and through this we can make a judgement that the avg ranking of “paid” apps are less than that of the “free” apps

4th Worksheet

Developer Insights: We have identified the top developers based on app rank and number of apps by using "developer_name" and "rank." and the top developers in the year of 2013 is “ActiveDevelopment” in all categories and lists.

5th Worksheet

Apps released by Developers over the course of year 
We can also observe that there was a spike in the number of apps released by developers in the year 2011 and the same is marked by the trend line and throughout “free” apps were released more in number than “paid” apps.